Add fast run for Arm machines (M1, M2, ...)

+++ Lib/Sources/UIUnitTestCLI/StartServerCommand.swift
@@ -10,38 +10,52 @@ struct StartServerCommand: AsyncParsableCommand {
     @Option
     var deviceIdentifier: String
     
+    @Flag
+    var forceInstall = false
+    
+    @Flag
+    var notPrebuildServer = false
+    
     mutating func run() async throws {
-//        let result: String = await executeShellCommand("xcrun simctl listapps booted")
-//
-//        let appInstalled = result.contains("bruno.mazzo.ServerUITests.xctrunner")
-//
-//        guard appInstalled else {
-//            throw UIUnitTestErrors.appNotInstalled
-//        }
-        
-        let serverRunnerZip = Bundle.module.url(forResource: "Server", withExtension: ".zip")!
-        
-        let tempDirectory = getTempFolder()
-        
-        let testRunnerZip = copyFile(file: serverRunnerZip, toFolder: tempDirectory)
-        
-        await executeShellCommand("unzip -o \(testRunnerZip.path) -d \(tempDirectory.relativePath)")
-        
-        let rootFolder = String(tempDirectory.pathComponents.joined(separator: "/").dropFirst())
-        
-        
-        await executeTestUntilServerStarts("""
-        xcodebuild -project \(rootFolder)/Server.xcodeproj \
-            -scheme ServerUITests -sdk iphonesimulator \
-            -destination "platform=iOS Simulator,id=\(deviceIdentifier)" \
-            test &
-        """)
-        
-//        -IDEBuildLocationStyle=Custom \
-//        -IDECustomBuildLocationType=Absolute \
-//        -IDECustomBuildProductsPath="$PWD/build/Products" \
-        
-//        await executeShellCommand("xcrun simctl launch booted bruno.mazzo.ServerUITests.xctrunner")
+        if isArmMac() && notPrebuildServer {
+            let result: String = await executeShellCommand("xcrun simctl listapps \(deviceIdentifier)")
+            
+            let appInstalled = result.contains("bruno.mazzo.ServerUITests.xctrunner")
+            
+            if !appInstalled || forceInstall {
+                let serverRunnerZip = Bundle.module.url(forResource: "PreBuild", withExtension: ".zip")!
+                
+                let tempDirectory = getTempFolder()
+                
+                let testRunnerZip = copyFile(file: serverRunnerZip, toFolder: tempDirectory)
+                
+                await executeShellCommand("unzip -o \(testRunnerZip.path) -d \(tempDirectory.relativePath)")
+                
+                let rootFolder = String(tempDirectory.pathComponents.joined(separator: "/").dropFirst())
+                
+                await executeShellCommand("xcrun simctl install \(deviceIdentifier) \(rootFolder)/ServerUITests-Runner.app")
+            }
+            
+            await executeShellCommand("xcrun simctl launch \(deviceIdentifier) bruno.mazzo.ServerUITests.xctrunner")
+        } else {
+            let serverRunnerZip = Bundle.module.url(forResource: "Server", withExtension: ".zip")!
+            
+            let tempDirectory = getTempFolder()
+            
+            let testRunnerZip = copyFile(file: serverRunnerZip, toFolder: tempDirectory)
+            
+            await executeShellCommand("unzip -o \(testRunnerZip.path) -d \(tempDirectory.relativePath)")
+            
+            let rootFolder = String(tempDirectory.pathComponents.joined(separator: "/").dropFirst())
+            
+            
+            await executeTestUntilServerStarts("""
+                xcodebuild -project \(rootFolder)/Server.xcodeproj \
+                    -scheme ServerUITests -sdk iphonesimulator \
+                    -destination "platform=iOS Simulator,id=\(deviceIdentifier)" \
+                    test &
+                """)
+        }
     }
     
     func getTempFolder() -> URL {
@@ -63,5 +77,20 @@ struct StartServerCommand: AsyncParsableCommand {
         
         try! FileManager.default.copyItem(at: initialPath, to: newPath)
         return newPath
+    }
+    
+    func isArmMac() -> Bool {
+        var systeminfo = utsname()
+        uname(&systeminfo)
+        let machine = withUnsafeBytes(of: &systeminfo.machine) {bufPtr->String in
+            let data = Data(bufPtr)
+            if let lastIndex = data.lastIndex(where: {$0 != 0}) {
+                return String(data: data[0...lastIndex], encoding: .isoLatin1)!
+            } else {
+                return String(data: data, encoding: .isoLatin1)!
+            }
+        }
+        print(machine)
+        return machine == "arm64"
     }
 }
